+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/events.cljs
@@ -232,13 +232,13 @@
 (reg-event-fx
   ::check-dct-later
   (fn [_ [_ job-id]]
-    {:dispatch-later
-     [{:ms 3000 :dispatch [::get-job-check-dct job-id]}]}))
+    {:dispatch-later [{:ms 3000 :dispatch [::get-job-check-dct job-id]}]}))
 
 (reg-event-fx
   ::check-dct
   (fn [_ [_ {:keys [id] :as _deployment}]]
-    {::cimi-api-fx/operation [id "check-dct" #(dispatch [::check-dct-later (:location %)])]}))
+    (let [on-success #(dispatch [::check-dct-later (:location %)])]
+      {::cimi-api-fx/operation [id "check-dct" on-success]})))
 
 ; What's the difference with the same event in deployment
 (reg-event-fx
@@ -262,36 +262,35 @@
   (fn [{{:keys [::spec/deployment] :as db} :db} [_ id first-step do-not-open-modal?]]
     (when (= :data first-step)
       (dispatch [::get-data-records]))
-    (let [body              (if (str/starts-with? id "module/")
-                              {:module {:href id}}
-                              {:deployment {:href id}})
+    (let [from-module?      (str/starts-with? id "module/")
           old-deployment-id (:id deployment)
-          on-success        #(dispatch [::get-deployment (:resource-id %)])]
+          on-success        #(dispatch [::get-deployment (:resource-id %)])
+          on-error          #(do
+                               (dispatch [::reset])
+                               (dispatch
+                                 [::messages-events/add
+                                  (let [{:keys [status message]} (response/parse-ex-info %)]
+                                    {:header  (cond-> "Error during creation of deployment"
+                                                      status (str " (" status ")"))
+                                     :content message
+                                     :type    :error})]))]
       (cond->
-        {:db               (assoc db ::spec/deployment nil
-                                     ::spec/selected-credential-id nil
-                                     ::spec/selected-infra-service nil
-                                     ::spec/deploy-modal-visible? (not (boolean do-not-open-modal?))
-                                     ::spec/active-step (or first-step :data)
-                                     ::spec/data-step-active? (= first-step :data)
-                                     ::spec/cloud-filter nil
-                                     ::spec/selected-cloud nil
-                                     ::spec/cloud-infra-services nil
-                                     ::spec/data-clouds nil
-                                     ::spec/license-accepted? false
-                                     ::spec/module-info nil
-                                     ::spec/selected-version nil
-                                     ::spec/original-module nil)
-         ::cimi-api-fx/add [:deployment body on-success
-                            :on-error #(do
-                                         (dispatch [::reset])
-                                         (dispatch
-                                           [::messages-events/add
-                                            (let [{:keys [status message]} (response/parse-ex-info %)]
-                                              {:header  (cond-> "Error during creation of deployment"
-                                                                status (str " (" status ")"))
-                                               :content message
-                                               :type    :error})]))]}
+        {:db (assoc db ::spec/deployment nil
+                       ::spec/selected-credential-id nil
+                       ::spec/selected-infra-service nil
+                       ::spec/deploy-modal-visible? (not (boolean do-not-open-modal?))
+                       ::spec/active-step (or first-step :data)
+                       ::spec/data-step-active? (= first-step :data)
+                       ::spec/cloud-filter nil
+                       ::spec/selected-cloud nil
+                       ::spec/cloud-infra-services nil
+                       ::spec/data-clouds nil
+                       ::spec/license-accepted? false
+                       ::spec/module-info nil
+                       ::spec/selected-version nil
+                       ::spec/original-module nil)}
+        from-module? (assoc ::cimi-api-fx/add [:deployment {:module {:href id}} on-success :on-error on-error])
+        (not from-module?) (assoc ::cimi-api-fx/operation [id "clone" on-success :on-error on-error])
         old-deployment-id (assoc ::cimi-api-fx/delete [old-deployment-id #() :on-error #()])))))
 
 (defn on-error-reselect-credential
@@ -375,13 +374,7 @@
 (reg-event-fx
   ::deployment-operation
   (fn [_ [_ id operation]]
-    (let [callback (fn [response]
-                     (if (instance? js/Error response)
-                       (do (dispatch [::set-error-message
-                                      (str "Error occured during \"" operation
-                                           "\" action on deployment")
-                                      (-> response response/parse-ex-info :message)])
-                           (dispatch [::set-submit-loading? false]))
+    (let [on-success (fn [response]
                        (let [{:keys [message]} (response/parse response)
                              success-msg {:header  (str operation " action called successfully")
                                           :content message
@@ -390,8 +383,14 @@
                          (dispatch [::messages-events/add success-msg])
                          (dispatch [::deployments-detail-events/get-deployment id])
                          (dispatch [::history-events/navigate
-                                    (str "deployment/" (general-utils/id->uuid id))]))))]
-      {::cimi-api-fx/operation [id operation callback]})))
+                                    (str "deployment/" (general-utils/id->uuid id))])))
+          on-error   (fn [response]
+                       (dispatch [::set-error-message
+                                  (str "Error occured during \"" operation
+                                       "\" action on deployment")
+                                  (-> response response/parse-ex-info :message)])
+                       (dispatch [::set-submit-loading? false]))]
+      {::cimi-api-fx/operation [id operation on-success :on-error on-error]})))
 
 (reg-event-fx
   ::edit-deployment