+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Dev tools - Integration of portal
 - Deployment modal - update button disabled when registries is added in module newer version bugfix #1038
 - Cloudflare preview deploys - Simulate behavior of nginx prod server
+- Deployment - clone bugfix #1045
+- CIMI-API - operation on-error default handler and signature change
+- General - operation signature change alignment
 
 ## [2.32.10] - 2022-12-06